+++ src/routes/oath.ts
@@ -151,7 +151,10 @@ export const HubApiCall:any = async function (myapifunc:Function,refresh_token:s
 	// first try to see if the api call goes through if it does then send response back
 	try
 	{
+
 		let result = await myapifunc()
+
+
 		// api call went through everything is fine
 		return result;
 	}
@@ -159,11 +162,13 @@ export const HubApiCall:any = async function (myapifunc:Function,refresh_token:s
 	{
 		try
 		{
+
 			//error in the api call get a new access token
 			let result = await getHubspotAccessToken(refresh_token);
 
 			if(result.status ==400)//result.response.status == 400)
 			{
+
 				if(process.env.NODE_ENV == 'test')
 				{
 					// return error result for the testing purpose
